Start roughing in task.evening
Having evening flags on tasks is an attribute I really liked in things.
It allows for 'today' to be separated into day/night tasks and help keep
day and night goals better separated.

There are still a few wrinkles to be sorted as far as how evening
works with sorting scopes.