Improve comments and fix redirect loop bug in calendar views.
After taking an action on a calendar source users would get trapped in
a redirect loop, as the 'referer' of the list *was* the list. This is
a drawback of the simple redirect based workflow I chose for this.
Refactoring the currently simple UI might be a better approach long term
though.