Add API token request handling.
This still needs a few more tests but it is generally in the right
shape. The mobile app will require a long lived API token. This is the
'most' secure option but I feel that docket is small enough and there
will be an API to managed ApiTokens as well.