Improve internal db state for tasks
Add and remove tasks to the view ID lists when tasks are created
this should fix tasks not showing up after creation. I've also made
local updates simpler for completed tasks. They are removed from the
local DB instead of updated in place. This might need to be revised
again when completed views are added or when undo is added for task
completion.