Get prototype working for app_auth
This gives the UX I was initially going for. I also don't seem to need
to deploy applinks for this which is wonderfull. There is a slightly
janky step where the user has to choose the app to handle the redirect.

This flow did yield a refresh and access token though which is ideal.