Get new today repository mostly working
This should help pave the way for infinite scroll for upcoming as we can
keep track of the 'end' and use that as the start for the next range.

Each day is then stored in the daily repo and can be read as one block
or fetched from network incrementally.