Big refactor of the the networking 'layer'
Introduce a few new abstractions as there were some clear patterns
in the code. Notably the error handling for decoding bytes into a string
and then into a map was repeated many times. Dart lends itself well to a
callback style pattern so I'm going to lean into that for this routine
task. This also allows error handling to be greatly improved so that in
in development and hopefully in sentry errors I can get much more
context.